ref: move runCommand to utils.ts
